pyproject.toml
[project]
name = "ai-assistant"
version = "0.1.0"
description = "Personal AI assistant with memory, tool use, and Claude API integration"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.42.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.0",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.7.0",
    "click>=8.1.0",
    "rich>=13.9.0",
    "prompt-toolkit>=3.0.48",
    "sqlite-utils>=3.37",
    "python-telegram-bot>=21.0",
    "apscheduler>=3.10",
    "watchdog>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[project.scripts]
assistant = "assistant.cli.app:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/assistant"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
